backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
email-validator==2.2.0

# Redis client (with async support)
redis==5.2.0

# Google APIs
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# Google OAuth authentication
authlib==1.3.2
PyJWT==2.9.0

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.12

# Image processing
Pillow==11.0.0

# Scheduler
APScheduler==3.10.4

# iCalendar generation
icalendar==6.1.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
fakeredis==2.26.2
httpx==0.27.2

